Update price models
