+++ tests/test_database.py
@@ -8,8 +8,7 @@ from sqlalchemy import event, inspect, text
 from sqlmodel import Session, SQLModel, create_engine, select
 
 from niveshpy.database import _iregexp, get_session, initialize
-from niveshpy.infrastructure.sqlite.models import Account, Security
-from niveshpy.models.price import Price
+from niveshpy.infrastructure.sqlite.models import Account, Price, Security
 from niveshpy.models.security import SecurityCategory, SecurityType
